+++ packages/vite-node/src/hmr/index.ts
@@ -1,3 +1,2 @@
 export * from './emitter'
-export * from './hotContext'
 export * from './hmr'
